deprecated/requirements.txt
pypika
requests
numpy
pyTelegramBotAPI
openai
python-dotenv
discord.py
aiohttp